+++ hc/front/tests/test_add_victorops.py
@@ -11,7 +11,7 @@ class AddVictorOpsTestCase(BaseTestCase):
     def test_instructions_work(self):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.get(self.url)
-        self.assertContains(r, "incident management system")
+        self.assertContains(r, "incident management platform")
 
     def test_it_works(self):
         form = {"value": "http://example.org"}